Adjusted how the creation of media works
Show and Path are no longer required during creation of a media.
I decided that forcing them to be required would be more trouble
than it was worth considering I'm going to be the only using the api.

Also cleaned up the media creation controller to allow it to be
more expandable and handle errors a little bit better. I was going
to use async like I did in the update controller but I decided that
there was no use for it at the moment.