Added Videoangular, site now tracks what video each user watches
Videoangular is now the video player used for presenting media. This is
mostly to make it look a bit nicer and more importantly allow me to
more easilty track user views. Each time a user watches a show, the
episode and show will be sent to two seperate arrays within the user
object. They will be inside a pill object that also attaches information
like their progress in the series and how far into the episode they got to.
These for now are 0 and will be fleshed out more later. There is also a
date_watched variable which will update the pill everytime the user watches
something more then once.

At the moment a lot of the logic used to handle all this pill stuff is
done on the client side, this is not desired and will be changed at a
later date. Doing it client side at first was only done to get a better
handle on how to divide the labor.

There was additonally some minor refactoring along the way, I'm  deleting a
bunch of comments I left in when I was first doing the MEAN stack tutorial.
Much of these comments are useless and describing basic syntax. More of
this will be done in a future commit.